Stop multiple deliveries of STOPSIGNAL
Under Ptrace, STOPSIGNAL is sent to the inferior threads once by the
superior thread and once by the ckpt-thread of the inferior. STOPSIGNAL
is blocked while the inferior thread is executing the signal handler and
so the signal is becomes pending and is delivered right after returning
from stopthisthread.

This commit fixes this by making ckpt-thread disable/re-enable signal
handler for STOPSIGNAL right before callback_sleep_between_ckpt().

git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1610 ba6dc0ae-232f-0410-883f-9ebac64fc565