Changes to when to inform coord of RUNNING state
When running with ptrace, the user threads go through an extra step
before returning to executing user code. We shouldn't inform coordinator
of RUNNING state untill all threads are ready to go to the user code.
Failing to do so, can result in a sequence where the coordinator sends a
kill message which can't be entertained by superior process because it
is still waiting to attach to the inferior process (which has already
processed the message and exited).

git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1770 ba6dc0ae-232f-0410-883f-9ebac64fc565