Implement Connection plugin as a true plugin
- The checkpoint image is just a memory dump of the process. All the
information required to recreate child processes and file descriptors is
now part of the process memory. This information used to be part of
"DMTCP-header" in the checkpoint image. Now this is part of the memory
dump.

- During restart, the child process are forked off only after the
process's memory has been restored. Once the child processes have been
forked, each process restores/recreates the file descriptors including
socket connections for which it was ckpt-leader. The file descriptors
are then passed to non-leader processes over UNIX domain sockets.

- Socketpair aren't properly restored at this time. Instead, they are
restored as AF_INET sockets.

- make check-screen is current failing due to a bug related to
packet-mode detection for pseudo-terminals.

git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1786 ba6dc0ae-232f-0410-883f-9ebac64fc565