Fix restart of distributed application. using getsockname for the listen socket will not return IP addres of the host. Instead "0.0.0.0" will be returned. This information was used at restart time for NS service and thus restore of connection to remote caused "Connection refused" error.
This fix assumes that communication with coordinator is performed through the same IP 
address (and thus Ethernet device) as all other connections.
On clusters this assumption may be incorrect. There may be service network that is ok
for coordinator communicatiuons and computational network (that is usually faster) that
is dedicated for parallel process communication.
In future we can provide an option that will provide preferred address for network interface
that should be used for communication.



git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1813 ba6dc0ae-232f-0410-883f-9ebac64fc565